add cast long

+++ maven-project-demo/src/main/java/org/example/AlgorithmsAndStructures/ChangeDigit.java
@@ -10,6 +10,6 @@ public class ChangeDigit {
         k = (long) Math.pow(10, k);
         double saved = num % k;
         num = num - (num % (k * 10L));
-        return ((long) (num + (newDigit * k) + saved));
+        return ((long) (num + ((long) newDigit * k) + saved));
     }
 }
